Change first_ to size_t and last_ to long long

+++ SJCVector.h
@@ -12,8 +12,8 @@
 class SJCVector {
 	int* ptr_;		//Class manages resource
 	size_t size_;
-	int first_;
-	int last_;
+	size_t first_;
+	long long last_;
 	std::string name_;
 
 public:
